Add ResourcePrintPreview component and update routing for print view
