+++ src/Routers/routes/ResourceRoutes.tsx
@@ -4,6 +4,7 @@ import BoardView from "@/components/Resource/ResourceBoard";
 import ResourceDetails from "@/components/Resource/ResourceDetails";
 import { ResourceDetailsUpdate } from "@/components/Resource/ResourceDetailsUpdate";
 import ListView from "@/components/Resource/ResourceList";
+import ResourcePrintPreview from "@/components/Resource/ResourcePrintPreview";
 
 import { AppRoutes } from "@/Routers/AppRouter";
 
@@ -16,6 +17,7 @@ const ResourceRoutes: AppRoutes = {
   "/resource/list": () => <ListView />,
   "/resource/:id": ({ id }) => <ResourceDetails id={id} />,
   "/resource/:id/update": ({ id }) => <ResourceDetailsUpdate id={id} />,
+  "/resource/:id/print": ({ id }) => <ResourcePrintPreview id={id} />,
 };
 
 export default ResourceRoutes;
